requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
python-dotenv==1.0.0
textmagic-rest==2.0.0
